refactor request signals and split signals to their apps
| Principle                   | Benefit                                                                               |
| --------------------------- | ------------------------------------------------------------------------------------- |
| **Separation of concerns**  | Keeps each app self-contained and responsible for its own behavior.                   |
| **Decoupling**              |  app becomes a pure data model (no cross-app logic).                   |
| **Easier testing**          | Each app’s tests only need to deal with its own signals.                              |
| **Scalability**             | Adding new apps or features doesn’t bloat a single .                      |
| **Maintainability**         | Team members working on one app don’t need to understand unrelated logic from others. |
| **Avoids circular imports** | Local signals inside apps can import Notification lazily or via .     |
