+++ feedit_app/notifications/apps.py
@@ -4,7 +4,3 @@ from django.apps import AppConfig
 class NotificationsConfig(AppConfig):
     default_auto_field = "django.db.models.BigAutoField"
     name = "notifications"
-
-    def ready(self):
-        """Import signals when the app is ready."""
-        __import__("notifications.signals")
