+++ feedit_app/reviews/apps.py
@@ -1,10 +1,10 @@
 from django.apps import AppConfig
 
 
-class RequestsConfig(AppConfig):
+class ReviewsConfig(AppConfig):
     default_auto_field = "django.db.models.BigAutoField"
-    name = "company_requests"
+    name = "reviews"
 
     def ready(self):
         """Import signals when the app is ready."""
-        __import__("company_requests.signals")
+        __import__("reviews.signals")
